Refactor project into a Cargo Workspace.
Moved server code into new server crate
Moved client code into new client crate
Moved common code from lib.rs into new common crate
Left top-level main and lib files to stash code to be re-used
