+++ hw05_client_server_msger/Cargo.toml
@@ -1,14 +1,7 @@
-[package]
-name = "hw05_client_server_msger"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+members = [ 
+    "client", 
+    "common",
+    "server"
+] 
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "client"
-path = "bin/client.rs"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
