+++ hw05_client_server_msger/client/src/main.rs
@@ -1,0 +1,5 @@
+fn main() {
+    println!("In client::main()");
+    client::run_client();
+    println!("Leaving client::main()");
+}
